this commit uses the python stdlib logging facility and removes the various logging output directories, replacing them with an info, error, debug logging level scheme with configuration file.  Default console logs only error and info.  Debug logs are directed to a file
