+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "program": "${file}",
             "console": "integratedTerminal",
-            "envFile": "${workspaceFolder}/src/.env",
+            "envFile": "/workspace/AutoPatch-LLM/src/fuzzing-service/.env",
         },
         {
             "name": "Python Debugger: AutoPatch Main",
