+++ src/fuzzing-service/.gitignore
@@ -1,0 +1,1 @@
+.env
