+++ src/fuzzing-service/config.json.example
@@ -1,0 +1,19 @@
+{
+    "version": "0.1.0",
+    "appname": "autopatch.fuzzing-service",
+    "logging_config" : "/workspace/AutoPatch-LLM/src/fuzzing-service/logging-config.json",
+    "fuzz_svc_input_codebase_path": "/workspace/AutoPatch-LLM/src/codebase",
+    "compiled_binary_executables_output_path": "/workspace/AutoPatch-LLM/src/executables",
+    "compiler_tool_full_path": "/usr/bin/gcc",
+    "fuzzer_tool_name": "afl",
+    "fuzzer_tool_version": "2.52b",
+    "afl_tool_full_path": "/workspace/AutoPatch-LLM/afl-2.52b/afl-fuzz",
+    "afl_tool_seed_input_path": "/workspace/AutoPatch-LLM/src/input",
+    "afl_tool_output_path": "/workspace/AutoPatch-LLM/src/output",
+    "afl_tool_compiled_binary_executables_output_path": "/workspace/AutoPatch-LLM/src/executables_afl",
+    "afl_compiler_tool_full_path": "/workspace/AutoPatch-LLM/afl-2.52b/afl-gcc",
+    "fuzz_svc_output_topic": "fuzz-svc/output",
+    "mqtt_host": "localhost",
+    "mqtt_port": 1883,
+    "debug": false
+}
