+++ src/fuzzing-service/pyproject.toml
@@ -1,0 +1,7 @@
+[project]
+name = "autopatch.fuzzing-service"
+version = "0.1.0"
+requires-python = ">=3.12.3"
+
+[tool.isort]
+profile = "black"
