--- src/fuzzing-svc/.env.dev
@@ -1,1 +1,0 @@
-FUZZ_SVC_CONFIG='/workspace/AutoPatch-LLM/src/fuzzing-svc/config.json'
