--- src/fuzzing-svc/.gitignore
@@ -1,2 +1,0 @@
-.env
-debug.env
