--- src/fuzzing-svc/Taskfile.yml
@@ -1,12 +1,0 @@
-version: '3'
-
-tasks:
-  run:
-    desc: run AutoPatch-FuzzingService
-    cmds:
-      - python ./fuzzing-service-main.py
-  
-  debug:
-    desc: debug AutoPatch-FuzzingService
-    cmds:
-      - python -m pdb ./fuzzing-service-main.py
