--- src/fuzzing-svc/config.json.example
@@ -1,19 +1,0 @@
-{
-    "mqtt_host": "localhost",
-    "mqtt_port": 1883,
-    "command_log_path": "/workspace/AutoPatch-LLM/src/command_log",
-    "fuzz_svc_output_path": "/workspace/AutoPatch-LLM/src/command_log",
-    "fuzz_svc_output_output_topic": "fuzz-svc/output",
-    "fuzz_svc_input_codebase_path": "/workspace/AutoPatch-LLM/src/codebase",
-    "compiler_tool_full_path": "/usr/bin/gcc",
-    "afl_tool_full_path": "/workspace/AutoPatch-LLM/afl-2.52b/afl-fuzz",
-    "afl_tool_seed_input_path": "/workspace/AutoPatch-LLM/src/input",
-    "afl_tool_output_topic": "afl/output",
-    "afl_tool_output_path_template": "/workspace/AutoPatch-LLM/src/output_",
-    "afl_tool_compiled_binary_executables_output_path": "/workspace/AutoPatch-LLM/src/executables_afl",
-    "afl_compiler_tool_full_path": "/workspace/AutoPatch-LLM/afl-2.52b/afl-gcc",
-    "address_sanitizer_output_topic": "asan/output",
-    "address_sanitizer_output_path": "/workspace/AutoPatch-LLM/src/asan_bugLog",
-    "compiled_binary_executables_output_path": "/workspace/AutoPatch-LLM/src/executables",
-    "debug": false
-}
